+++ frontend/src/pages/HomePage/components/ToolsSection/FeatureSections/index.ts
